New alerts page base using buefy table prepared to the backend pagination
