+++ package.json
@@ -14,7 +14,7 @@
     "animate.css": "^4.1.1",
     "axios": "^0.24.0",
     "babel-plugin-transform-require-context": "^0.1.1",
-    "buefy": "^0.9.13",
+    "buefy": "^0.9.21",
     "chart.js": "^2.9.4",
     "core-js": "^3.19.1",
     "d3": "^7.1.1",
