+++ src/views/alerts/Alerts.vue
@@ -5,213 +5,27 @@
     :centerCol="9"
     v-if="isMounted"
   >
-    <b-loading
-      slot="loading"
-      :is-full-page="false"
-      v-model="isLoading"
-      :can-cancel="false"
-    />
-
     <AlertsFilters slot="left">
       <Loading :isLoading="loadingTableStatus" v-if="firstLoad" />
     </AlertsFilters>
 
-    <FullTable
-      slot="center"
-      :placeholder="$t('menu.alerts')"
-      :keys="keys"
-      :tableData="getAlerts"
-      class="table-alerts"
-      @pageRows="
-        (vals) => {
-          currentPageSelection = vals
-        }
-      "
-      @isPageChanged="handleClearAllSelections"
-      @clickedRow="handleClickedRow"
-      isClickable
-      :isLoadingTable="loadingTableStatus"
-    >
-      <template slot="customTopHeader">
-        <div
-          v-if="isCurrentPageSelected || selectedRows.length > 0"
-          style="margin-right: auto"
-        >
-          <b-button
-            type="is-primary"
-            size="is-small"
-            icon-pack="fas"
-            icon-left="check-circle"
-            class="has-text-weight-semibold mr-3 ml-3"
-            @click="handleMarkAsRead"
-          >
-            {{ $t('views.alerts.markRead') }}
-          </b-button>
-
-          <span
-            class="is-size-7"
-            v-if="!isAllPagesSelected && isCurrentPageSelected"
-          >
-            <span class="px-2">
-              {{ $t('views.alerts.selected', [currentPageSelection.length]) }}
-            </span>
-            <a @click="handleSelectAllPagesRows">
-              {{ $t('views.alerts.selectAll', [getAlerts.length]) }}
-            </a>
-          </span>
-
-          <span
-            class="is-size-7"
-            v-if="isAllPagesSelected && isCurrentPageSelected"
-          >
-            <span class="px-2">
-              {{ $t('views.alerts.allAlerts', [getAlerts.length]) }}
-            </span>
-            <a @click="handleClearAllSelections">
-              {{ $t('views.alerts.clearAll') }}
-            </a>
-          </span>
-        </div>
-
-        <div v-if="alerts.params.category" style="padding-left: 20px">
-          <b-button
-            type="is-primary"
-            size="is-small"
-            class="has-text-weight-semibold mr-3"
-            @click="removeParams"
-          >
-            {{ $t('views.alerts.showAll') }}
-          </b-button>
-        </div>
-
-        <RefreshButton tooltipMsg="Update Alerts Data" />
-      </template>
-
-      <template slot="headData">
-        <th style="width: 5%">
-          <div v-if="showCheckbox">
-            <b-checkbox
-              v-model="isCurrentPageSelected"
-              @input="handleSelectPageRows"
-            />
-          </div>
-        </th>
-        <v-th style="width: 10%" sortKey="alertCategory"> Type </v-th>
-        <v-th style="width: 10%" sortKey="date">
-          {{ $t('common.collumns.date') }}
-        </v-th>
-        <v-th style="width: 5%" :sortKey="alertSeveritySort">
-          {{ $t('common.collumns.severity') }}
-        </v-th>
-        <v-th style="width: 20%" sortKey="hostname">
-          {{ $t('common.collumns.hostname') }}
-        </v-th>
-        <v-th style="width: 10%" sortKey="alertCode">
-          {{ $t('common.collumns.code') }}
-        </v-th>
-        <v-th style="width: 40%" sortKey="description">
-          {{ $t('common.collumns.description') }}
-        </v-th>
-      </template>
-
-      <template slot="bodyData" slot-scope="rowData">
-        <td
-          class="py-0 px-0"
-          v-if="
-            rowData.scope.alertCategory !== 'AGENT' &&
-            rowData.scope.alertStatus === 'NEW'
-          "
-        >
-          <b-checkbox
-            v-model="rowData.scope.isChecked"
-            @input="
-              handleSelectRows(rowData.scope.isChecked, rowData.scope._id)
-            "
-            class="is-flex is-justify-content-center"
-            style="height: 32px"
-          />
-        </td>
-        <td v-else></td>
-        <TdContent :value="rowData.scope.alertCategory" />
-        <TdContent :value="rowData.scope.date" dataType="date" />
-        <TdIcon
-          :value="resolveIcon(rowData.scope.alertSeverity)"
-          @click.native="handleClickedRow([rowData.scope])"
-        />
-        <HostLink
-          v-if="rowData.scope.alertStatus !== 'DISMISSED'"
-          :hostname="rowData.scope.hostname ? rowData.scope.hostname : '-'"
-        />
-        <TdContent v-else :value="rowData.scope.hostname" />
-        <TdContent :value="rowData.scope.alertCode" />
-
-        <TdContent
-          :value="rowData.scope.description"
-          v-if="rowData.scope.description.length < 100"
-        />
-        <td v-if="rowData.scope.description.length > 100">
-          <b-icon
-            v-tooltip="options($t('common.general.fullDesc'))"
-            type="is-primary"
-            class="delete-icon"
-            pack="fas"
-            icon="file-alt"
-            @click.native="descriptionAlert(rowData.scope)"
-          />
-          {{ rowData.scope.description }}
-        </td>
-      </template>
-
-      <ExportButton
-        slot="export"
-        :url="`alerts?status=${alertStatus}`"
-        expName="alerts-data"
-      />
-    </FullTable>
+    <AlertsTable slot="center" />
   </ToggleColumns>
 </template>
 
 <script>
-import _ from 'lodash'
-import { bus } from '@/helpers/eventBus.js'
-import { mapGetters, mapActions, mapState, mapMutations } from 'vuex'
-import { descriptionAlertDialog } from '@/helpers/alertsDescDialog.js'
-import { resolveSeverityIcon } from '@/helpers/helpers.js'
-import paginationMixin from '@/mixins/paginationMixin.js'
-import TooltipMixin from '@/mixins/tooltipMixin.js'
-import localFiltersMixin from '@/mixins/localFiltersMixin.js'
-import hostnameLinkRow from '@/mixins/hostnameLinkRow.js'
+import { mapGetters, mapActions, mapMutations } from 'vuex'
 import ToggleColumns from '@/components/common/ToggleColumns.vue'
-import FullTable from '@/components/common/Table/FullTable.vue'
-import ExportButton from '@/components/common/ExportButton.vue'
-import TdContent from '@/components/common/Table/TdContent.vue'
-import TdIcon from '@/components/common/Table/TDIcon.vue'
-import HostLink from '@/components/common/Table/HostLink.vue'
-import AlertsFilters from '@/components/alerts/AlertsFilters.vue'
+import AlertsFilters from '@/components/alerts/Filters.vue'
+import AlertsTable from '@/components/alerts/Table.vue'
 import Loading from '@/components/common/Loading.vue'
-import RefreshButton from '@/components/common/RefreshButton.vue'
-
-const checkOrUncheck = (list, status, handleSelectRows) => {
-  _.map(list, (val) => {
-    if (val.alertCategory !== 'AGENT') {
-      val.isChecked = status
-      return handleSelectRows(val.isChecked, val._id)
-    }
-  })
-}
 
 export default {
-  mixins: [paginationMixin, localFiltersMixin, hostnameLinkRow, TooltipMixin],
   components: {
     ToggleColumns,
-    FullTable,
-    ExportButton,
-    TdContent,
-    TdIcon,
-    HostLink,
     AlertsFilters,
+    AlertsTable,
     Loading,
-    RefreshButton,
   },
   data() {
     return {
@@ -223,10 +37,6 @@ export default {
         'description',
         'alertSeverity',
       ],
-      selectedRows: [],
-      isCurrentPageSelected: false,
-      currentPageSelection: [],
-      isAllPagesSelected: false,
       isLoading: false,
       isMounted: false,
       alertStatus: 'NEW',
@@ -234,123 +44,19 @@ export default {
     }
   },
   async beforeMount() {
-    await this.getAlertsData({ status: this.alertStatus }).then(() => {
-      bus.$emit('data', this.getAlerts)
-      this.firstLoad = false
-    })
+    await this.getAlertsData({ status: this.alertStatus })
   },
   mounted() {
     this.isMounted = true
   },
   methods: {
-    ...mapActions(['getAlertsData', 'markAsReadAlertsPage']),
+    ...mapActions(['getAlertsData']),
     ...mapMutations(['SET_ALERTS_PARAMS']),
-    resolveIcon(value) {
-      return resolveSeverityIcon(value)
-    },
-    handleMarkAsRead() {
-      this.isLoading = true
-      const idList = _.uniqWith(this.selectedRows, _.isEqual)
-      this.markAsReadAlertsPage({ idList }).then(() => {
-        this.getAlerts
-        this.isCurrentPageSelected = false
-        this.isLoading = false
-        this.selectedRows = []
-      })
-    },
-    handleSelectRows(status, id) {
-      if (status) {
-        this.selectedRows.push(id)
-      } else {
-        this.selectedRows = _.filter(this.selectedRows, (val) => {
-          return val !== id
-        })
-      }
-    },
-    handleSelectPageRows(checked) {
-      checked
-        ? checkOrUncheck(this.currentPageSelection, true, this.handleSelectRows)
-        : checkOrUncheck(
-            this.currentPageSelection,
-            false,
-            this.handleSelectRows
-          )
-    },
-    handleSelectAllPagesRows() {
-      this.isAllPagesSelected = true
-      checkOrUncheck(
-        this.getAlerts,
-        this.isAllPagesSelected,
-        this.handleSelectRows
-      )
-    },
-    handleClearAllSelections() {
-      this.isAllPagesSelected = false
-      checkOrUncheck(
-        this.getAlerts,
-        this.isAllPagesSelected,
-        this.handleSelectRows
-      )
-    },
-    alertSeveritySort(row) {
-      return row.alertSeverity.length
-    },
-    removeParams() {
-      bus.$emit('onResetAction')
-    },
-    descriptionAlert(info) {
-      const data = {
-        code: info.alertCode,
-        host: info.hostname,
-        categ: info.alertCategory,
-        date: info.date,
-        desc: info.description,
-        severity: info.alertSeverity,
-      }
-      descriptionAlertDialog(data)
-    },
   },
   computed: {
-    ...mapState(['alerts']),
-    ...mapGetters(['getAlerts', 'showCheckbox', 'loadingTableStatus']),
-  },
-  watch: {
-    selectedRows(value) {
-      if (value.length < this.perPage) {
-        this.isCurrentPageSelected = false
-      } else {
-        this.isCurrentPageSelected = true
-      }
-    },
+    ...mapGetters(['loadingTableStatus']),
   },
 }
 </script>
 
-<style lang="scss">
-.table-alerts {
-  thead {
-    tr {
-      th {
-        &:first-child {
-          text-align: center !important;
-        }
-        &:last-child {
-          text-align: left !important;
-        }
-      }
-    }
-  }
-  tbody {
-    tr {
-      td {
-        &:first-child {
-          text-align: center;
-        }
-        &:last-child {
-          text-align: left;
-        }
-      }
-    }
-  }
-}
-</style>
+<style lang="scss"></style>
